Added a little CSS to login page.
Updated login middleware so that it redirects unauthorized away
from more routes.

Added client-side error checking for the login and signup forms.

Updated the login and signup views so that there is now a div
that is hidden whenever there are no errors. These errors can
either be created from the client or sent in response from the server.

Added getByEmail function to check if the email is already taken
when a user tries to sign up.

Also made sure that form data isn't lost when there is an error by
sending it back through handlebars.