Created a fetchStockPrice function that makes requests to Yahoo Finance's API to update the stock prices of the industries. We store only the most recent so that we can calculate return on stock holdings on the wallet page.
On app.js, I created an interval that calls the fetchStockPrice function,
as well as the calculatePortfolioValue function to each user so that
it updates automatically over time.

The returns on each stock holding is now calculated on the fly
and rendered onto the wallet.handlebars.

I also made a users.getAll() function that is used in the interval.

I also called xss on the routes in routes/wallet.js since that is
required.